Fix test directory structure
