+++ tests/pyathena/arrow/test_async_cursor.py
@@ -13,7 +13,7 @@ from pyathena.error import NotSupportedError, ProgrammingError
 from pyathena.model import AthenaQueryExecution
 from pyathena.result_set import AthenaResultSet
 from tests import ENV
-from tests.conftest import connect
+from tests.pyathena.conftest import connect
 
 
 class TestAsyncarrowCursor:
