+++ tests/pyathena/arrow/test_util.py
