+++ tests/pyathena/conftest.py
@@ -2,27 +2,13 @@
 import contextlib
 from io import BytesIO
 from pathlib import Path
-from urllib.parse import quote_plus
 
 import boto3
 import pytest
 import sqlalchemy
-from sqlalchemy.testing.plugin.pytestplugin import *  # noqa
-from sqlalchemy.testing.plugin.pytestplugin import (
-    pytest_sessionfinish as sqlalchemy_pytest_sessionfinish,
-)
-from sqlalchemy.testing.plugin.pytestplugin import (
-    pytest_sessionstart as sqlalchemy_pytest_sessionstart,
-)
-from sqlalchemy.testing.provision import (
-    configure_follower,
-    create_db,
-    drop_db,
-    temp_table_keyword_args,
-)
 
 from tests import ENV, SQLALCHEMY_CONNECTION_STRING
-from tests.util import read_query
+from tests.pyathena.util import read_query
 
 
 def pytest_sessionstart(session):
@@ -32,19 +18,6 @@ def pytest_sessionstart(session):
             _create_database(cursor)
             _create_table(cursor)
 
-    conn_str = (
-        SQLALCHEMY_CONNECTION_STRING + "&tblproperties=" + quote_plus("'table_type'='ICEBERG'")
-    )
-    session.config.option.dburi = [
-        conn_str.format(
-            region_name=ENV.region_name,
-            schema_name=ENV.schema,
-            s3_staging_dir=ENV.s3_staging_dir,
-            location=ENV.s3_staging_dir,
-        )
-    ]
-    sqlalchemy_pytest_sessionstart(session)
-
 
 def pytest_sessionfinish(session):
     with contextlib.closing(connect()) as conn:
@@ -52,12 +25,10 @@ def pytest_sessionfinish(session):
             _drop_database(cursor)
     _delete_rows()
 
-    sqlalchemy_pytest_sessionfinish(session)
-
 
 def _upload_rows():
     client = boto3.client("s3")
-    rows = Path(__file__).parent.resolve() / "resources" / "rows"
+    rows = Path(__file__).parents[1].resolve() / "resources" / "rows"
     for row in rows.iterdir():
         key = f"{ENV.s3_staging_key}{ENV.schema}/{row.stem}/{row.name}"
         client.upload_file(str(row), ENV.s3_staging_bucket, key)
@@ -70,7 +41,7 @@ def _upload_rows():
 
 def _delete_rows():
     client = boto3.client("s3")
-    rows = Path(__file__).parent.resolve() / "resources" / "rows"
+    rows = Path(__file__).parents[1].resolve() / "resources" / "rows"
     for row in rows.iterdir():
         key = f"{ENV.s3_staging_key}{ENV.schema}/{row.stem}/{row.name}"
         client.delete_object(Bucket=ENV.s3_staging_bucket, Key=key)
@@ -92,39 +63,6 @@ def _create_table(cursor):
         "create_table.sql.jinja2", s3_staging_dir=ENV.s3_staging_dir, schema=ENV.schema
     ):
         cursor.execute(q)
-
-
-@create_db.for_db("awsathena")
-def _awsathena_create_db(cfg, eng, ident):
-    with eng.begin() as conn:
-        try:
-            _awsathena_drop_db(cfg, conn, ident)
-        except Exception:
-            pass
-
-    with eng.begin() as conn:
-        conn.exec_driver_sql(f"CREATE DATABASE {ident}")
-        conn.exec_driver_sql(f"CREATE DATABASE {ident}_test_schema")
-        conn.exec_driver_sql(f"CREATE DATABASE {ident}_test_schema_2")
-
-
-@drop_db.for_db("awsathena")
-def _awsathena_drop_db(cfg, eng, ident):
-    with eng.begin() as conn:
-        conn.exec_driver_sql(f"DROP DATABASE {ident} CASCADE")
-        conn.exec_driver_sql(f"DROP DATABASE {ident}_test_schema CASCADE")
-        conn.exec_driver_sql(f"DROP DATABASE {ident}_test_schema_2 CASCADE")
-
-
-@configure_follower.for_db("awsathena")
-def _awsathena_configure_follower(config, ident):
-    config.test_schema = f"{ident}_test_schema"
-    config.test_schema_2 = f"{ident}_test_schema_2"
-
-
-@temp_table_keyword_args.for_db("awsathena")
-def _awsathena_temp_table_keyword_args(cfg, eng):
-    return {"prefixes": ["TEMPORARY"]}
 
 
 def connect(schema_name="default", **kwargs):
