+++ tests/pyathena/fastparquet/__init__.py
