+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 isolated_build = true
-envlist = py38,py39,py310,py311
+envlist = py{38,39,310,311},py{38,39,310,311}-sqla
 
 [gh-actions]
 python =
@@ -16,8 +16,13 @@ allowlist_externals =
 commands =
     poetry install -v
     make chk
-    poetry run pytest -n 8 --cov pyathena --cov-report html --cov-report term
+    poetry run pytest -n 8 --cov pyathena --cov-report html --cov-report term tests/pyathena/
 passenv =
     AWS_*
     TOXENV
     GITHUB_*
+
+[testenv:py{38,39,310,311}-sqla]
+commands =
+    poetry install -v
+    poetry run pytest -n 8 --cov pyathena --cov-report html --cov-report term tests/sqlalchemy/
